requirements.txt
aiogram>=3.4,<4.0
SQLAlchemy>=2.0,<2.1
greenlet>=3.0,<4.0
asyncpg>=0.29,<1.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
APScheduler>=3.10,<4.0
gspread>=6.1,<7.0
google-auth>=2.29,<3.0
openpyxl>=3.1,<4.0
python-dateutil>=2.9,<3.0
pytest>=8.1,<9.0
pytest-asyncio>=0.23,<0.24
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
python-multipart>=0.0.9
jinja2>=3.1,<4.0
itsdangerous>=2.1,<3.0
alembic>=1.13,<2.0
